% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_fourier_features.R
\name{random_fourier_features}
\alias{random_fourier_features}
\title{Generate random Fourier features}
\usage{
random_fourier_features(x, w = NULL, b = NULL, num_f = NULL,
  sigma = NULL, seed = NULL)
}
\arguments{
\item{x}{Random variable x.}

\item{w}{Random coefficients.}

\item{b}{Random offsets.}

\item{num_f}{Number of random Fourier features.}

\item{sigma}{Smooth parameter of RBF kernel.}

\item{seed}{The seed for controlling random number generation. Use if you want to replicate results exactly. Default is NULL.}
}
\value{
A list containing the p-value \code{p} and statistic \code{Sta}
}
\description{
Generate random Fourier features
}
\examples{
random_fourier_features(rnorm(1000),num_f=25)
}

