% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Repeat copies of a matrix or a vector just like MATLAB's repmat function}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{a matrix or a vector}

\item{m}{Number of row repeats}

\item{n}{Number of column repeats}
}
\value{
A matrix with associated row and column repeats
}
\description{
Repeat copies of a matrix or a vector just like MATLAB's repmat function
}
\examples{
repmat(matrix(rnorm(2*2),ncol=2),m=2,n=3)
}

