% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIT.R
\name{RCIT}
\alias{RCIT}
\title{RCIT and RCoT - tests whether x and y are conditionally independent given z. Calls RIT if z is empty.}
\usage{
RCIT(x, y, z = NULL, approx = "hbe", corr = TRUE, num_f = 25,
  seed = NULL)
}
\arguments{
\item{x}{Random variable x.}

\item{y}{Random variable y.}

\item{z}{Random variable z.}

\item{approx}{Method for approximating the null distribution. Default is the "lpd4," the Lindsay-Pilla-Basak method. Other options include "gamma" for the Satterthwaite-Welch method, "hbe" for the Hall-Buckley-Eagleson method, and "chi2" for a normalized statistic.}

\item{corr}{TRUE corresponds to RCoT and FALSE to RCIT. Default is FALSE.}

\item{num_f}{Number of features for conditioning set. Default is 25.}

\item{seed}{The seed for controlling random number generation. Use if you want to replicate results exactly. Default is NULL.}
}
\value{
A list containing the p-value \code{p} and statistic \code{Sta}
}
\description{
RCIT and RCoT - tests whether x and y are conditionally independent given z. Calls RIT if z is empty.
}
\examples{
RCIT(rnorm(1000),rnorm(1000),rnorm(1000));

x=rnorm(10000);
y=(x+rnorm(10000))^2;
z=rnorm(10000);
RCIT(x,y,z,seed=2);
}

